git-fzf, gh-fzf: Implement dynamic fzf header
Implements functions that display fzf header using a bash function.
The header is printed before fzf's input and the '--header-lines' option
is used to correctly display the header.

The header shows current modes of git search and ripgrep mode toggles.

As a side-effect, toggling modes now not only refreshes the preview,
but reloads fzf so that new header can be loaded.